+++ src/__tests__/client-navigation/__fixtures__/pages/a.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
-import Link from 'next/link';
 import { useRouter } from 'next/router';
 import { basename } from 'path';
+import { PageA } from '../PageA';
 
 const expectedRoute = `/${basename(__filename, '.tsx')}`;
 
 export default function ClientSideNavigationA() {
-  const { replace, route } = useRouter();
+  const { route } = useRouter();
 
   // Ensure client side navigation updates router & page in same tick
   if (route !== expectedRoute) {
@@ -15,30 +15,10 @@ export default function ClientSideNavigationA() {
     );
   }
 
-  const href = '/b?foo=bar';
-  const hrefObject = { pathname: '/b', query: { foo: 'bar' } };
-
-  const goToPageBstring = () => {
-    replace(href);
-  };
-
-  const goToPageBObject = () => {
-    replace(hrefObject);
-  };
-
   return (
-    <div>
-      <h2>This is page A</h2>
-
-      <Link href={href}>
-        <a>Go to B with Link (with string)</a>
-      </Link>
-      <Link href={hrefObject}>
-        <a>Go to B with Link (with object)</a>
-      </Link>
-
-      <a onClick={goToPageBstring}>Go to B programmatically (with string)</a>
-      <a onClick={goToPageBObject}>Go to B programmatically (with object)</a>
-    </div>
+    <PageA
+      href="/b?foo=bar"
+      hrefObject={{ pathname: '/b', query: { foo: 'bar' } }}
+    />
   );
 }
